requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
boto3==1.35.0
yt-dlp==2024.9.27
supabase==2.7.4
weasyprint==62.3
httpx==0.27.2
python-multipart==0.0.9
